requirements.txt
django==3.2
ifcopenshell==0.7.0.230818
pydantic
ezdxf
djangorestframework
celery
eventlet;sys_platform=="win32"
attrs
cattrs
lark==1.1.8
django-grappelli==2.15.6
python-dotenv==1.0.0
psycopg2==2.9.9
numpy
python-fcl==0.7.0.5
docxtpl==0.16.7
docutils
black
gunicorn;sys_platform!="win32"
uvicorn;sys_platform!="win32"
